fix response and bug no photo was selected but error

+++ features/attendance/data/model.go
@@ -14,7 +14,7 @@ type Attendance struct {
 	AttendanceDate   string
 	Attendance       string
 	AttendanceStatus string
-	WorkTime         int
+	WorkTime         float32
 	ClockInLocation  string
 	ClockOutLocation string
 	ClockInOSM       string
